pyproject.toml
[project]
name = "magi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.21.0",
    "python-dotenv>=1.2.1",
    "python-frontmatter>=1.1.0",
]
